requirements.txt
fastapi
uvicorn
pydantic
python-multipart
tortoise-orm
aiomysql
aerich
python-jose[cryptography]
passlib[bcrypt]
sqlglot[rs]
loguru
sqlparse
tqdm
nltk
aiofiles
func_timeout
numpy
pandas